Removing all unnecessary files
