--- packages/cli/src/args.ts
@@ -1,5 +1,0 @@
-export const pathArg = {
-  name: 'paths',
-  description:
-    'The paths or globs that checkup will operate on. If no paths or globs are provided, checkup will run on the entire directory beginning at --cwd.',
-};
