--- packages/cli/src/index.ts
@@ -1,4 +1,0 @@
-export { run } from '@oclif/command';
-export { runCommand } from './run-command';
-export { getReporter } from './reporters/get-reporter2';
-export { default as Checkup } from './commands/run';
