+++ packages/cli/src/reporters/available-task-reporter.ts
@@ -1,17 +1,13 @@
-import { ui } from '@checkup/core';
-import { bold } from 'chalk';
-import TaskList from '../task-list';
-
-export function reportAvailableTasks(pluginTasks: TaskList) {
-  ui.blankLine();
-  ui.log(bold.white('AVAILABLE TASKS'));
-  ui.blankLine();
-  if (pluginTasks.size > 0) {
-    pluginTasks.fullyQualifiedTaskNames.forEach((taskName) => {
-      ui.log(`  ${taskName}`);
+export function reportAvailableTasks(availableTasks: string[]) {
+  console.log();
+  console.log('AVAILABLE TASKS');
+  console.log();
+  if (availableTasks.length > 0) {
+    availableTasks.forEach((taskName) => {
+      console.log(`  ${taskName}`);
     });
   } else {
-    ui.log(`  No tasks found`);
+    console.log(`  No tasks found`);
   }
-  ui.blankLine();
+  console.log();
 }
