--- packages/cli/src/reporters/available-task-reporter2.ts
@@ -1,13 +1,0 @@
-export function reportAvailableTasks(availableTasks: string[]) {
-  console.log();
-  console.log('AVAILABLE TASKS');
-  console.log();
-  if (availableTasks.length > 0) {
-    availableTasks.forEach((taskName) => {
-      console.log(`  ${taskName}`);
-    });
-  } else {
-    console.log(`  No tasks found`);
-  }
-  console.log();
-}
