+++ packages/cli/src/reporters/console-reporter.ts
@@ -1,31 +1,40 @@
-import { CheckupMetadata, RunFlags, ui } from '@checkup/core';
+import { CheckupMetadata, ui } from '@checkup/core';
 import { Log } from 'sarif';
 import { renderActions, renderCLIInfo, renderInfo } from './reporter-utils';
 import { writeOutputFile } from './sarif-file-writer';
 import { success } from 'log-symbols';
+import { ReportOptions } from './get-reporter';
 
-export function report(result: Log, flags?: RunFlags) {
-  let { cwd } = flags!;
-  let { rules } = result.runs[0].tool.driver;
-  let metaData = result.properties as CheckupMetadata;
+export default class ConsoleReporter {
+  options: ReportOptions;
 
-  renderInfo(metaData);
+  constructor(options: ReportOptions) {
+    this.options = options;
+  }
 
-  ui.log('Checkup ran the following task(s) successfully:');
-  rules!
-    .map((rule) => rule.id)
-    .sort()
-    .forEach((taskName) => {
-      ui.log(success, taskName);
-    });
+  report(result: Log) {
+    let { cwd } = this.options;
+    let { rules } = result.runs[0].tool.driver;
+    let metaData = result.properties as CheckupMetadata;
 
-  ui.blankLine();
-  ui.log('Results have been saved to the following file:');
-  writeOutputFile('{default}', cwd, result);
+    renderInfo(metaData);
 
-  renderActions(result.properties?.actions);
+    ui.log('Checkup ran the following task(s) successfully:');
+    rules!
+      .map((rule) => rule.id)
+      .sort()
+      .forEach((taskName) => {
+        ui.log(success, taskName);
+      });
 
-  ui.blankLine();
+    ui.blankLine();
+    ui.log('Results have been saved to the following file:');
+    writeOutputFile('{default}', cwd, result);
 
-  renderCLIInfo(metaData);
+    renderActions(result.properties?.actions);
+
+    ui.blankLine();
+
+    renderCLIInfo(metaData);
+  }
 }
