+++ packages/cli/src/reporters/json-reporter.ts
@@ -1,11 +1,20 @@
 import { Log } from 'sarif';
-import { ui, RunFlags } from '@checkup/core';
+import { ui } from '@checkup/core';
 import { writeOutputFile } from './sarif-file-writer';
+import { ReportOptions } from './get-reporter';
 
-export function report(result: Log, flags?: RunFlags) {
-  if (flags && flags['output-file']) {
-    writeOutputFile(flags['output-file'], flags.cwd, result);
-  } else {
-    ui.styledJSON(result);
+export default class JsonReporter {
+  options: ReportOptions;
+
+  constructor(options: ReportOptions) {
+    this.options = options;
+  }
+
+  report(result: Log) {
+    if (this.options.outputFile) {
+      writeOutputFile(this.options.outputFile, this.options.cwd, result);
+    } else {
+      ui.styledJSON(result);
+    }
   }
 }
